requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
mysql-connector-python==8.2.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
scipy==1.11.4
fuzzywuzzy==0.18.0
python-levenshtein==0.23.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4